Add SettingsStore class for managing application settings
- Implemented SettingsStore class with methods to load, save, and provide default settings.
- Added functionality to read settings from a configuration file and validate them.
- Introduced helper functions for parsing and converting configuration values.
- Created a platform-independent method to determine the configuration directory.
- Ensured settings are saved in a structured format with appropriate error handling.

+++ src/app/application.cpp
@@ -373,7 +373,8 @@ bool Application::create_workers() {
                 }
 
                 if (sent) {
-                    int ret = sendto(sock, reinterpret_cast<const char*>(buffer.ptr()), buffer.size, 0,
+                    int ret = sendto(sock, reinterpret_cast<const char*>(buffer.ptr()),
+                                   static_cast<int>(buffer.size), 0,
                                    reinterpret_cast<struct sockaddr*>(&sin), sizeof(sin));
                         if (ret > 0) {
                             g_total_packets++;
